+++ exmol/exmol.py
@@ -1,22 +1,26 @@
-from rdkit.Chem import rdFMCS as MCS
-import requests
-import numpy as np
-from sklearn.cluster import DBSCAN
-from sklearn.decomposition import PCA
-import selfies as sf
+from typing import *
+
 import itertools
-import math, os
+import math
+import requests  # type: ignore
+import numpy as np
+import matplotlib.pyplot as plt  # type: ignore
+import matplotlib as mpl  # type: ignore
+import selfies as sf  # type: ignore
+import tqdm  # type: ignore
+
+from ratelimit import limits, sleep_and_retry  # type: ignore
+from sklearn.cluster import DBSCAN  # type: ignore
+from sklearn.decomposition import PCA  # type: ignore
+from rdkit.Chem import MolFromSmiles as smi2mol  # type: ignore
+from rdkit.Chem import MolToSmiles as mol2smi  # type: ignore
+from rdkit.Chem import rdchem  # type: ignore
+from rdkit.Chem.Draw import MolToImage as mol2img  # type: ignore
+from rdkit.Chem import rdFMCS as MCS  # type: ignore
+
+
 from . import stoned
 from .plot_utils import _mol_images, _image_scatter
-from rdkit.Chem import MolFromSmiles as smi2mol
-from rdkit.Chem.Draw import MolToImage as mol2img
-from rdkit.Chem import MACCSkeys
-import rdkit.Chem
-import matplotlib.pyplot as plt
-import matplotlib as mpl
-from typing import *
-import tqdm
-from ratelimit import limits, sleep_and_retry
 from .data import *
 
 _calculator = None
@@ -25,8 +29,7 @@ _calculator = None
 def _fp_dist_matrix(smiles, fp_type, _pbar):
     mols = [(smi2mol(s), _pbar.update(0.5))[0] for s in smiles]
     # Sorry about the one-line. Just sneaky insertion of progressbar update
-    fp = [(stoned.get_fingerprint(m, fp_type), _pbar.update(0.5))[0]
-          for m in mols]
+    fp = [(stoned.get_fingerprint(m, fp_type), _pbar.update(0.5))[0] for m in mols]
     # 1 - Ts because we want distance
     dist = list(
         1 - stoned.TanimotoSimilarity(x, y) for x, y in itertools.product(fp, repeat=2)
@@ -103,7 +106,7 @@ def get_basic_alphabet() -> Set[str]:
     to_remove.extend(["[P]", "[#P]", "[=P]"])
 
     a -= set(to_remove)
-    a.add("[O-1expl]")
+    a.add("[O-1]")
     return a
 
 
@@ -114,17 +117,17 @@ def run_stoned(
     max_mutations: int = 2,
     min_mutations: int = 1,
     alphabet: Union[List[str], Set[str]] = None,
-    _pbar: Any = None
-) -> Tuple[List[str], List[str]]:
+    _pbar: Any = None,
+) -> Tuple[List[str], List[float]]:
     """Run ths STONED SELFIES algorithm. Typically not used, call :func:`sample_space` instead.
 
     :param s: SMILES string to start from
     :param fp_type: Fingerprint type
-    :param num_samples: Number of molecules to generate per mutation
+    :param num_samples: Number of total molecules to generate
     :param max_mutations: Maximum number of mutations
     :param min_mutations: Minimum number of mutations
     :param alphabet: Alphabet to use for mutations, typically from :func:`get_basic_alphabet()`
-    :return: SMILES and SELFIES generated
+    :return: SMILES and SCORES generated
     """
     if alphabet is None:
         alphabet = list(sf.get_semantic_robust_alphabet())
@@ -136,19 +139,20 @@ def run_stoned(
     if mol == None:
         raise Exception("Invalid starting structure encountered")
 
+    # want it so after sampling have num_samples
     randomized_smile_orderings = [
-        stoned.randomize_smiles(mol) for _ in range(num_samples)
+        stoned.randomize_smiles(mol) for _ in range(num_samples // len(num_mutation_ls))
     ]
 
     # Convert all the molecules to SELFIES
     selfies_ls = [sf.encoder(x) for x in randomized_smile_orderings]
 
-    all_smiles_collect = []
-    all_selfies_collect = []
+    all_smiles_collect: List[str] = []
+    all_selfies_collect: List[str] = []
     for num_mutations in num_mutation_ls:
         # Mutate the SELFIES:
         if _pbar:
-            _pbar.set_description(f'🥌STONED🥌 Mutations: {num_mutations}')
+            _pbar.set_description(f"🥌STONED🥌 Mutations: {num_mutations}")
         selfies_mut = stoned.get_mutated_SELFIES(
             selfies_ls.copy(), num_mutations=num_mutations, alphabet=alphabet
         )
@@ -161,7 +165,7 @@ def run_stoned(
 
     # Work on:  all_smiles_collect
     if _pbar:
-        _pbar.set_description(f'🥌STONED🥌 Done')
+        _pbar.set_description(f"🥌STONED🥌 Done")
     canon_smi_ls = []
     for item in all_smiles_collect:
         mol, smi_canon, did_convert = stoned.sanitize_smiles(item)
@@ -189,39 +193,42 @@ def run_chemed(
     origin_smiles: str,
     num_samples: int,
     similarity: float = 0.1,
-    fp_type: str = 'ECFP4',
+    fp_type: str = "ECFP4",
     _pbar: Any = None,
 ) -> Tuple[List[str], List[float]]:
     """
     This method is similar to STONED but works by quering PubChem
+
     :param origin_smiles: Base SMILES
     :param num_samples: Minimum number of returned molecules. May return less due to network timeout or exhausting tree
     :param similarity: Tanimoto similarity to use in query (float between 0 to 1)
-    :return: SMILES
+    :param fp_type: Fingerprint type
+    :return: SMILES and SCORES
     """
     if _pbar:
-        _pbar.set_description('⚡CHEMED⚡ is Experimental ☠️')
+        _pbar.set_description("⚡CHEMED⚡")
     else:
-        print('⚡CHEMED⚡ is Experimental ☠️')
-    url = f'https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/fastsimilarity_2d/smiles/{origin_smiles}/property/CanonicalSMILES/JSON'
+        print("⚡CHEMED⚡")
+    url = f"https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/fastsimilarity_2d/smiles/{requests.utils.quote(origin_smiles)}/property/CanonicalSMILES/JSON"
     try:
-        reply = requests.get(url, params={
-            'Threshold': int(similarity), 'MaxRecords': num_samples},
-            headers={'accept': 'text/json'},
-            timeout=10)
+        reply = requests.get(
+            url,
+            params={"Threshold": int(similarity * 100), "MaxRecords": num_samples},
+            headers={"accept": "text/json"},
+            timeout=10,
+        )
     except requests.exceptions.Timeout:
-        print('Pubchem seems to be down right now ️☠️☠️')
+        print("Pubchem seems to be down right now ️☠️☠️")
         return [], []
     try:
         data = reply.json()
     except:
         return [], []
-    smiles = [d['CanonicalSMILES']
-              for d in data['PropertyTable']['Properties']]
-    smiles = set(smiles)
+    smiles = [d["CanonicalSMILES"] for d in data["PropertyTable"]["Properties"]]
+    smiles = list(set(smiles))
 
     if _pbar:
-        _pbar.set_description(f'Received {len(smiles)} similar molecules')
+        _pbar.set_description(f"Received {len(smiles)} similar molecules")
 
     mol0 = smi2mol(origin_smiles)
     mols = [smi2mol(s) for s in smiles]
@@ -239,28 +246,70 @@ def run_chemed(
     return smiles, scores
 
 
+def run_custom(
+    origin_smiles: str,
+    data: List[Union[str, rdchem.Mol]],
+    fp_type: str = "ECFP4",
+    _pbar: Any = None,
+    **kwargs,
+) -> Tuple[List[str], List[float]]:
+    """
+    This method is similar to STONED but uses a custom dataset provided by the user
+
+    :param origin_smiles: Base SMILES
+    :param data: List of SMILES or RDKit molecules
+    :param fp_type: Fingerprint type
+    :return: SMILES and SCORES
+    """
+    if _pbar:
+        _pbar.set_description("⚡CUSTOM⚡")
+    else:
+        print("⚡CUSTOM⚡")
+    mol0 = smi2mol(origin_smiles)
+    fp0 = stoned.get_fingerprint(mol0, fp_type)
+    scores = []
+    smiles = []
+    # drop invalid molecules
+    for d in data:
+        if isinstance(d, str):
+            m = smi2mol(d)
+        else:
+            m = d
+        if m is None:
+            continue
+        smiles.append(mol2smi(m))
+        fp = stoned.get_fingerprint(m, fp_type)
+        scores.append(stoned.TanimotoSimilarity(fp0, fp))
+        if _pbar:
+            _pbar.update()
+    return smiles, scores
+
+
 def sample_space(
     origin_smiles: str,
     f: Union[
-        Callable[[str, str], float], Callable[[
-            List[str], List[str]], List[float]]
+        Callable[[List[str], List[str]], List[float]],
+        Callable[[List[str], List[str]], List[float]],
     ],
     batched: bool = True,
     preset: str = "medium",
+    data: List[Union[str, rdchem.Mol]] = None,
     method_kwargs: Dict = None,
     num_samples: int = None,
-    stoned_kwargs: Dict = None
+    stoned_kwargs: Dict = None,
 ) -> List[Example]:
     """Sample chemical space around given SMILES
 
-    This will evaluate the given function and run the :func:`run_stoned` function over chemical space around molecule.
+    This will evaluate the given function and run the :func:`run_stoned` function over chemical space around molecule. ``num_samples`` will be
+    set to 3,000 by default if using STONED and 150 if using ``chemed``.
 
     :param origin_smiles: starting SMILES
     :param f: A function which takes in SMILES and SELFIES and returns predicted value. Assumed to work with lists of SMILES/SELFIES unless `batched = False`
     :param batched: If `f` is batched
-    :param preset: Can be wide, medium, or narrow. Determines how far across chemical space is sampled. Try `"chemed"` experimental preset to only sample commerically available compounds.
-    :param method_kwargs: More control over STONED or CHEMED can be set here. See :func:`run_stoned` and :func:`run_chemed`
-    :param num_samples: Number of desired samples. Can be set in `method_kwargs` (overrides) or here. `None` means default from preset.
+    :param preset: Can be wide, medium, or narrow. Determines how far across chemical space is sampled. Try `"chemed"` preset to only sample commerically available compounds.
+    :param data: If not None and preset is `"custom"` will use this data instead of generating new ones.
+    :param method_kwargs: More control over STONED, CHEMED and CUSTOM can be set here. See :func:`run_stoned`, :func:`run_chemed` and  :func:`run_custom`
+    :param num_samples: Number of desired samples. Can be set in `method_kwargs` (overrides) or here. `None` means default for preset
     :param stoned_kwargs: Backwards compatible alias for `methods_kwargs`
     :return: List of generated :obj:`Example`
     """
@@ -273,7 +322,7 @@ def sample_space(
     origin_smiles = stoned.sanitize_smiles(origin_smiles)[1]
     if origin_smiles is None:
         raise ValueError("Given SMILES does not appear to be valid")
-    smi_yhat = batched_f([origin_smiles], [sf.encoder(origin_smiles)])
+    smi_yhat = np.asarray(batched_f([origin_smiles], [sf.encoder(origin_smiles)]))
     try:
         iter(smi_yhat)
     except TypeError:
@@ -286,7 +335,7 @@ def sample_space(
     if method_kwargs is None:
         method_kwargs = {}
         if preset == "medium":
-            method_kwargs["num_samples"] = 1500 if num_samples is None else num_samples
+            method_kwargs["num_samples"] = 3000 if num_samples is None else num_samples
             method_kwargs["max_mutations"] = 2
             method_kwargs["alphabet"] = get_basic_alphabet()
         elif preset == "narrow":
@@ -294,11 +343,13 @@ def sample_space(
             method_kwargs["max_mutations"] = 1
             method_kwargs["alphabet"] = get_basic_alphabet()
         elif preset == "wide":
-            method_kwargs["num_samples"] = 600 if num_samples is None else num_samples
+            method_kwargs["num_samples"] = 3000 if num_samples is None else num_samples
             method_kwargs["max_mutations"] = 5
             method_kwargs["alphabet"] = sf.get_semantic_robust_alphabet()
         elif preset == "chemed":
             method_kwargs["num_samples"] = 150 if num_samples is None else num_samples
+        elif preset == "custom" and data is not None:
+            method_kwargs["num_samples"] = len(data)
         else:
             raise ValueError(f'Unknown preset "{preset}"')
     try:
@@ -313,49 +364,54 @@ def sample_space(
     # STONED
     if preset.startswith("chem"):
         smiles, scores = run_chemed(origin_smiles, _pbar=pbar, **method_kwargs)
+    elif preset == "custom":
+        smiles, scores = run_custom(
+            origin_smiles, data=cast(Any, data), _pbar=pbar, **method_kwargs
+        )
     else:
         smiles, scores = run_stoned(origin_smiles, _pbar=pbar, **method_kwargs)
     selfies = [sf.encoder(s) for s in smiles]
 
-    pbar.set_description('😀Calling your model function😀')
+    pbar.set_description("😀Calling your model function😀")
     fxn_values = batched_f(smiles, selfies)
 
     # pack them into data structure with filtering out identical
     # and nan
     exps = [
         Example(
-            smiles=origin_smiles,
-            selfies=sf.encoder(origin_smiles),
-            similarity=1.0,
-            yhat=smi_yhat,
+            origin_smiles,
+            sf.encoder(origin_smiles),
+            1.0,
+            cast(Any, smi_yhat),
             index=0,
             is_origin=True,
         )
     ] + [
-        Example(sm, se, s, np.squeeze(y), index=0)
+        Example(sm, se, s, cast(Any, np.squeeze(y)), index=0)
         for i, (sm, se, s, y) in enumerate(zip(smiles, selfies, scores, fxn_values))
         if s < 1.0 and np.isfinite(np.squeeze(y))
     ]
-    for i, e in enumerate(exps):
-        e.index = i
+
+    for i, e in enumerate(exps):  # type: ignore
+        e.index = i  # type: ignore
 
     pbar.reset(len(exps))
-    pbar.set_description('🔭Projecting...🔭')
+    pbar.set_description("🔭Projecting...🔭")
 
     # compute distance matrix
     full_dmat = _fp_dist_matrix(
         [e.smiles for e in exps],
         method_kwargs["fp_type"] if ("fp_type" in method_kwargs) else "ECFP4",
-        _pbar=pbar
+        _pbar=pbar,
     )
 
-    pbar.set_description('🥰Finishing up🥰')
+    pbar.set_description("🥰Finishing up🥰")
 
     # compute PCA
     pca = PCA(n_components=2)
     proj_dmat = pca.fit_transform(full_dmat)
-    for e in exps:
-        e.position = proj_dmat[e.index, :]
+    for e in exps:  # type: ignore
+        e.position = proj_dmat[e.index, :]  # type: ignore
 
     # do clustering everwhere (maybe do counter/same separately?)
     # clustering = AgglomerativeClustering(
@@ -363,10 +419,10 @@ def sample_space(
     # Just do it on projected so it looks prettier.
     clustering = DBSCAN(eps=0.15, min_samples=5).fit(proj_dmat)
 
-    for i, e in enumerate(exps):
-        e.cluster = clustering.labels_[i]
+    for i, e in enumerate(exps):  # type: ignore
+        e.cluster = clustering.labels_[i]  # type: ignore
 
-    pbar.set_description('🤘Done🤘')
+    pbar.set_description("🤘Done🤘")
     pbar.close()
     return exps
 
@@ -386,15 +442,13 @@ def _select_examples(cond, examples, nmols):
             result.append(close_counter)
 
     # trim, in case we had too many cluster
-    result = sorted(result, key=lambda v: v.similarity *
-                    cond(v), reverse=True)[:nmols]
+    result = sorted(result, key=lambda v: v.similarity * cond(v), reverse=True)[:nmols]
 
     # fill in remaining
     ncount = sum([cond(e) for e in result])
     fill = max(0, nmols - ncount)
     result.extend(
-        sorted(examples, key=lambda v: v.similarity *
-               cond(v), reverse=True)[:fill]
+        sorted(examples, key=lambda v: v.similarity * cond(v), reverse=True)[:fill]
     )
 
     return list(filter(cond, result))
@@ -442,7 +496,7 @@ def lime_explain(examples: List[Example], descriptor_type: str) -> np.ndarray:
 
 
 def cf_explain(examples: List[Example], nmols: int = 3) -> List[Example]:
-    """From given :obj:`Examples`, find best counterfactuals using :ref:`readme_link:counterfactual generation`
+    """From given :obj:`Examples<Example>`, find closest counterfactuals (see :doc:`index`)
 
     :param examples: Output from :func:`sample_space`
     :param nmols: Desired number of molecules
@@ -460,11 +514,10 @@ def cf_explain(examples: List[Example], nmols: int = 3) -> List[Example]:
 
 def rcf_explain(
     examples: List[Example],
-    delta: Union[float, Tuple[float, float]] = (-1, 1),
+    delta: Union[Any, Tuple[float, float]] = (-1, 1),
     nmols: int = 4,
 ) -> List[Example]:
-    """From given :obj:`Examples`, find best counterfactuals using :ref:`readme_link:counterfactual generation`
-
+    """From given :obj:`Examples<Example>`, find closest counterfactuals (see :doc:`index`)
     This version works with regression, so that a counterfactual is if the given example is higher or
     lower than base.
 
@@ -482,14 +535,12 @@ def rcf_explain(
         return e.yhat + delta[1] <= examples[0].yhat
 
     hresult = (
-        [] if delta[0] is None else _select_examples(
-            is_high, examples[1:], nmols // 2)
+        [] if delta[0] is None else _select_examples(is_high, examples[1:], nmols // 2)
     )
     for i, h in enumerate(hresult):
         h.label = f"Increase ({i+1})"
     lresult = (
-        [] if delta[1] is None else _select_examples(
-            is_low, examples[1:], nmols // 2)
+        [] if delta[1] is None else _select_examples(is_low, examples[1:], nmols // 2)
     )
     for i, l in enumerate(lresult):
         l.label = f"Decrease ({i+1})"
@@ -506,7 +557,7 @@ def plot_space(
     offset: int = 0,
     ax: Any = None,
     cartoon: bool = False,
-    rasterized: bool = False
+    rasterized: bool = False,
 ):
     """Plot chemical space around example and annotate given examples.
 
@@ -536,29 +587,15 @@ def plot_space(
         cmap = "Accent"
 
     else:
-        colors = [e.yhat for e in examples]
+        colors = cast(Any, [e.yhat for e in examples])
         normalizer = plt.Normalize(min(colors), max(colors))
         cmap = "viridis"
     space_x = [e.position[0] for e in examples]
     space_y = [e.position[1] for e in examples]
     if cartoon:
         # plot shading, lines, front
-        ax.scatter(
-            space_x,
-            space_y,
-            50,
-            "0.0",
-            lw=2,
-            rasterized=rasterized
-        )
-        ax.scatter(
-            space_x,
-            space_y,
-            50,
-            "1.0",
-            lw=0,
-            rasterized=rasterized
-        )
+        ax.scatter(space_x, space_y, 50, "0.0", lw=2, rasterized=rasterized)
+        ax.scatter(space_x, space_y, 50, "1.0", lw=0, rasterized=rasterized)
         ax.scatter(
             space_x,
             space_y,
@@ -567,7 +604,7 @@ def plot_space(
             cmap=cmap,
             lw=2,
             alpha=0.1,
-            rasterized=rasterized
+            rasterized=rasterized,
         )
     else:
         ax.scatter(
@@ -577,14 +614,13 @@ def plot_space(
             cmap=cmap,
             alpha=0.5,
             edgecolors="none",
-            rasterized=rasterized
+            rasterized=rasterized,
         )
     # now plot cfs/annotated points
     ax.scatter(
         [e.position[0] for e in exps],
         [e.position[1] for e in exps],
-        c=normalizer(
-            [e.cluster if highlight_clusters else e.yhat for e in exps]),
+        c=normalizer([e.cluster if highlight_clusters else e.yhat for e in exps]),
         cmap=cmap,
         edgecolors="black",
     )
@@ -596,10 +632,10 @@ def plot_space(
     for e in exps:
         if not e.is_origin:
             titles.append(f"Similarity = {e.similarity:.2f}\n{e.label}")
-            colors.append(base_color)
+            colors.append(cast(Any, base_color))
         else:
             titles.append("Base")
-            colors.append(base_color)
+            colors.append(cast(Any, base_color))
     _image_scatter(x, y, imgs, titles, colors, ax, offset=offset)
     ax.axis("off")
     ax.set_aspect("auto")
@@ -616,7 +652,7 @@ def plot_cf(
 ):
     """Draw the given set of Examples in a grid
 
-    :param exps: Small list of :obj:Example which will be drawn
+    :param exps: Small list of :obj:`Example` which will be drawn
     :param fig: Figure to plot onto
     :param figure_kwargs: kwargs to pass to :func:`plt.figure<matplotlib.pyplot.figure>`
     :param mol_size: size of rdkit molecule rendering, in pixles
@@ -635,7 +671,7 @@ def plot_cf(
         C = math.ceil(len(imgs) / R)
     if fig is None:
         if figure_kwargs is None:
-            figure_kwargs = {'figsize': (12, 8)}
+            figure_kwargs = {"figsize": (12, 8)}
         fig, axs = plt.subplots(R, C, **figure_kwargs)
     else:
         axs = fig.subplots(R, C)
@@ -644,7 +680,7 @@ def plot_cf(
         title = "Base" if e.is_origin else f"Similarity = {e.similarity:.2f}\n{e.label}"
         title += f"\nf(x) = {e.yhat:.3f}"
         axs[i].set_title(title)
-        axs[i].imshow(np.asarray(img), gid=f'rdkit-img-{i}')
+        axs[i].imshow(np.asarray(img), gid=f"rdkit-img-{i}")
         axs[i].axis("off")
     for j in range(i, C * R):
         axs[j].axis("off")