+++ docs/source/changelog.rst
@@ -1,6 +1,29 @@
 Change Log
 ==========
 
+v1.0.0 (2022-1-21)
+-------------------
+* Switched to SELFIES v2.0 release, which has a different token list
+* Removed "experimental" tag from Chemed and Custom methods
+* Type annotations now pass ``mypy``
+* Paper models now generate SVGs correctly and fixed token issues
+
+v0.6.0 (2022-1-17)
+-------------------
+* Changed behavior of num_samples so that it is not affected by mutation count in STONED
+
+v0.5.2 (2022-1-4)
+-------------------
+* Fixed SMILES escaping in URL in chemed
+
+v0.5.1 (2021-11-16)
+-------------------
+* Fixed similarity float vs int in chemed
+
+v0.5.0 (2021-10-26)
+-------------------
+* Added custom lists for counterfactual source (contributed by @maclandrol)
+
 v0.4.1 (2021-09-27)
 -------------------
 * Fixed randint usage in stoned.py